Control.Monad.Fraxl

The Fraxl Monad

type FreerT f = FreeT (Ap f)

Fraxl is based on a particular Freer monad. This Freer monad has applicative optimization, which is used to parallelize effects.

type Fraxl r = FreerT (Union r)

Fraxl is just the FreerT monad transformer, applied with Union. This is because Fraxl is just a free monad over a variety of data sources.

type Fetch f m a = ASeq f a -> m (ASeq m a)

Data sources produce Fetch functions. They take a sequence of effeects as an argument, and return a corresponding sequence of monadic actions which are used to wait on the results.

class Monad m => DataSource f m where

A data source is an effect f that operates in some monad m. Given a sequence of effects, a data source should use m to prepare a corresponding sequence of results.

Methods

fetch :: Fetch f m q

runFraxl :: forall f m a. DataSource f m => FreerT f m a -> m a

Runs a Fraxl computation. This takes FreerT as a parameter rather than Fraxl, because Fraxl is meant for a union of effects, but it should be possible to run a singleton effect.

simpleAsyncFetch :: MonadIO m => (forall x. f x -> IO x) -> Fetch f m a

A simple method of turning an IO bound computation into a DataSource Fetch.

The Sequence of Effects

data ASeq f a :: (* -> *) -> * -> * where

Constructors

ANil :: ASeq f ()

ACons :: f a1 -> ASeq f u -> ASeq f (a1, u)

reduceASeq :: Applicative f => ASeq f u -> f u

reduceASeq a sequence of applicative effects into an applicative.

hoistASeq :: (forall x. f x -> g x) -> ASeq f a -> ASeq g a

Transform a sequence of f into a sequence of g.

traverseASeq :: Applicative h => (forall x. f x -> h (g x)) -> ASeq f a -> h (ASeq g a)

Traverse a sequence with resepect to its interpretation type f.

rebaseASeq :: ASeq f u -> (forall x. (x -> y) -> ASeq f x -> z) -> (v -> u -> y) -> ASeq f v -> z

It may not look like it, but this appends two sequences. See Dave Menendez's work for more explanation.

Caching

newtype CachedFetch f a

Caching in Fraxl works by translating FreerT f into FreerT (CachedFetch f), then running with CachedFetch's DataSource. That instance requires f to to have a GCompare instance.

The CachedFetch instance uses a MonadState to track cached requests. The state variable is a DMap from the 'dependent-map' package. Keys are requests, and values are MVars of the results.

Constructors

CachedFetch (f a)

runCachedFraxl :: forall m f a. (MonadIO m, DataSource f m, GCompare f) => FreerT f m a -> DMap f MVar -> m (a, DMap f MVar)

Runs a Fraxl computation with caching using a given starting cache. Alongside the result, it returns the final cache.

evalCachedFraxl :: forall m f a. (MonadIO m, DataSource f m, GCompare f) => FreerT f m a -> m a

Like runCachedFraxl, except it starts with an empty cache and discards the final cache.

module Data.GADT.Compare

Fraxl Monads

class Monad m => MonadFraxl f m where

Class for Fraxl-capable monads.

Methods

dataFetch :: f a -> m a

dataFetch is used to make a request of type f.