Control.Monad.Fraxl

Description

Fraxl is a library based on Facebook's Haxl. The goal is to decompose Haxl into more general parts, in order to form a stronger composition with better type safety and purity.

Usage

Using Fraxl is fairly similar to Haxl. You define a request data type (often a GADT), and a DataSource instance. With this, Fraxl is able to perform requests concurrently.

data MySource a where
  MyString :: MySource String
  MyInt :: MySource Int

class Monad m => DataSource f m
  fetch :: ASeq f a -> m (ASeq m a)

instance MonadIO m => DataSource MySource m where
  fetch ANil = return ANil
  fetch (ACons f fs) = (ACons . liftIO . wait) $ liftIO (async $ downloadSource f) * fetch fs

You'll notice a few things here. For one, a data source can choose what monad it lives in. Unlike Haxl, which only lets you live in IO, Fraxl is a monad transformer, allowing you to use arbitrary underlying monads. Thus, maintaining state between fetches can be left up to the data source. This can be used for several things, such as caching or session management.

ASeq :: (* -> *) -> * -> * is similar to a heterogenous list. It is the data structure used by the fast free applicative. Interpreting this is akin to interpreting the free applicative.

The fetch method takes a list of f requests, and for each request, returns an m action that waits on the response. That is, fetch should start background threads for requests, and return all the actions for Fraxl to block with until they complete. This way, Fraxl can have many requests start their work in parallel, and call all their wait-actions together.

Composition

Fraxl is a composition of general tools. At the base of this composition is a free monad transformer ([the basis of which is described here](http:/elvishjerricco.github.io20160408/applicative-effects-in-free-monads.html)). This is because Fraxl (and Haxl) is necessarily a free monad. It's taking arbitrary data sources of kind * -> *, and constructing a monad out of them. Since there exists a free monad transformer with applicative optimization, there's no reason not to use it and get the transformer structure for free.

NOTE: This free monad is not in the latest version of free yet. So free is used as a git submodule to gain access to it.

The next layer of the composition is the free applicative. The free monad with applicative optimization uses any applicative (rather than any functor, as with the traditional free monad). Since the free applicative uses any type of kind * -> *, it is the perfect candidate for this layer. It allows Fraxl to see all the requests made in an applicative computation at once, which is how Fraxl can parallelize them.

The final layer is the data source layer. It is user-specified, but will often be a dependent open union. The dependent open union is essentially a nested either type over any number of types. If all of those types are data sources, the dependent open union allows Fraxl to handle all of them as one data source, in one layer of Fraxl. The nice thing about this is that it makes it type safe to use a data source. Whereas Haxl will simply trust that you know what you're doing, Fraxl will make it a type error to forget to initialize a data source, or call a computation without guaranteeing its data source is available.

The data source layer can be easily modified. Caching is a substitution of this layer that replaces the data source with one that caches the results of the original. It does this with a dependent map, whose keys are requests, and whose values are MVars of the results. If an uncached request is requested, an empty MVar is inserted into the cache map, the original fetch is called, and the result is stored in the MVar. If a cached request is requested, the wait-action returned will simply be readMVar.

The Fraxl Monad

type FreerT f = FreeT (Ap f)

Fraxl is based on a particular Freer monad. This Freer monad has applicative optimization, which is used to parallelize effects.

type Fraxl r = FreerT (Union r)

Fraxl is just the FreerT monad transformer, applied with Union. This is because Fraxl is just a free monad over a variety of data sources.

type Fetch f m a = ASeq f a -> m (ASeq m a)

Data sources produce Fetch functions. They take a sequence of effeects as an argument, and return a corresponding sequence of monadic actions which are used to wait on the results.

class Monad m => DataSource f m where

A data source is an effect f that operates in some monad m. Given a sequence of effects, a data source should use m to prepare a corresponding sequence of results.

Methods

fetch :: Fetch f m q

runFraxl :: forall f m a. DataSource f m => FreerT f m a -> m a

Runs a Fraxl computation. This takes FreerT as a parameter rather than Fraxl, because Fraxl is meant for a union of effects, but it should be possible to run a singleton effect.

simpleAsyncFetch :: MonadIO m => (forall x. f x -> IO x) -> Fetch f m a

A simple method of turning an IO bound computation into a DataSource Fetch.

The Sequence of Effects

data ASeq f a :: (* -> *) -> * -> * where

Constructors

ANil :: ASeq f ()

ACons :: f a1 -> ASeq f u -> ASeq f (a1, u)

reduceASeq :: Applicative f => ASeq f u -> f u

reduceASeq a sequence of applicative effects into an applicative.

hoistASeq :: (forall x. f x -> g x) -> ASeq f a -> ASeq g a

Transform a sequence of f into a sequence of g.

traverseASeq :: Applicative h => (forall x. f x -> h (g x)) -> ASeq f a -> h (ASeq g a)

Traverse a sequence with resepect to its interpretation type f.

rebaseASeq :: ASeq f u -> (forall x. (x -> y) -> ASeq f x -> z) -> (v -> u -> y) -> ASeq f v -> z

It may not look like it, but this appends two sequences. See Dave Menendez's work for more explanation.

Caching

newtype CachedFetch f a

Caching in Fraxl works by translating FreerT f into FreerT (CachedFetch f), then running with CachedFetch's DataSource. That instance requires f to to have a GCompare instance.

The CachedFetch instance uses a MonadState to track cached requests. The state variable is a DMap from the 'dependent-map' package. Keys are requests, and values are MVars of the results.

Constructors

CachedFetch (f a)

runCachedFraxl :: forall m f a. (MonadIO m, DataSource f m, GCompare f) => FreerT f m a -> DMap f MVar -> m (a, DMap f MVar)

Runs a Fraxl computation with caching using a given starting cache. Alongside the result, it returns the final cache.

evalCachedFraxl :: forall m f a. (MonadIO m, DataSource f m, GCompare f) => FreerT f m a -> m a

Like runCachedFraxl, except it starts with an empty cache and discards the final cache.

module Data.GADT.Compare

Fraxl Monads

class Monad m => MonadFraxl f m where

Class for Fraxl-capable monads.

Methods

dataFetch :: f a -> m a

dataFetch is used to make a request of type f.